PERF: address review feedback on GH-64429
- Hoist HAS_PYARROW and get_option imports to module top per
  jbrockmendel review
- Drop the Python validation loop (_is_list_of_str_or_na) and replace
  with an O(1) first-element gate + pa.array try/except for mixed-type
  lists. Removes the per-element Python iteration that scaled with
  input size and let validation be done in pyarrow C
- Move the SeriesConstructionFromListOfStrings asv from strings.py to
  ctors.py alongside SeriesConstructors
